Safari-609.1.11.1/LayoutTests/fast/workers/worker-crash-with-invalid-location-expected.txt

Blocked access to external URL http://example.com/worker.js
CONSOLE MESSAGE: line 18: Cannot load http://example.com/worker.js due to access control checks.
Test worker fetch of blocked url. Should print a "PASS" statement.

PASS
